fix: handle godot build irregularities
It has now been several releases that have had assets pulled or rebuilt
after the fact, with modified checksums. This change also forces
verification of checksums even when they are present in the checksum
file, to be additionally certain that users will not encounter
mismatches when using gdvm.

Also includes minor refactors to logging in the script.